+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.9.12] 2021
+### Features
+- Arch
+    - AUR: new custom action to quickly reinstall a package
+    <p align="center">
+            <img src="https://raw.githubusercontent.com/vinifmor/bauh-files/master/pictures/releases/0.9.12/aur_reinstall.png">
+    </p>
+    
 ### Improvements
 - Arch
     - repositories/AUR search time (=~ -70%)